Authentication and markdown editor

+++ app/Http/Controllers/ThreadController.php
@@ -7,6 +7,12 @@ use Illuminate\Http\Request;
 
 class ThreadController extends Controller
 {
+    function __construct()
+    {
+        return $this->middleware('auth')->except('index');
+    }
+    
+    
     /**
      * Display a listing of the resource.
      *
@@ -45,8 +51,7 @@ class ThreadController extends Controller
         ]);
 
         //store
-
-        Thread::create($request->all());
+        auth()->user()->threads()->create($request->all());
 
         //redirect
         return back()->withMessage('Thread Created!');
@@ -83,12 +88,17 @@ class ThreadController extends Controller
      */
     public function update(Request $request, Thread $thread)
     {
+        if(auth()->user()->id !== $thread->user_id){
+            abort(401,"unauthorized");
+        }
+
         //validate
         $this->validate($request, [
             'subject' => 'required|min:10',
             'type'    => 'required',
             'thread'  => 'required|min:20'
         ]);
+
 
         $thread->update($request->all());
 
@@ -104,6 +114,9 @@ class ThreadController extends Controller
      */
     public function destroy(Thread $thread)
     {
+        if(auth()->user()->id !== $thread->user_id){
+            abort(401,"unauthorized");
+        }
         $thread->delete();
 
         return redirect()->route('thread.index')->withMessage("Thread Deleted!");
